requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
pytest
awscli
flake8
python-dotenv>=0.5.1

matplotlib>=3.2.2
imbalanced_learn==0.7.0
numpy>=1.18.5
pandas_profiling>=2.9.0
scipy>=1.5.0
sklearn_evaluation>=0.5
plotly>=4.9.0
joblib>=0.16.0
pandas>=1.0.5
imblearn==0.0
scikit_learn==1.1.1
